- Added Elastic aggregation for patent ownership evolution - finished

+++ src/main/java/kiv/zcu/knowledgeipr/core/model/search/chartquery/PatentOwnershipEvolutionQuery.java
@@ -11,7 +11,7 @@ import java.util.List;
  * Manages the creation of a search returning evolution of patents ownership by a specified company (owner)
  * in a specified category
  */
-public class PatentOwnershipEvolutionQuery extends ChartQuery<Integer, Integer> {
+public class PatentOwnershipEvolutionQuery extends ChartQuery<Long, Long> {
 
     private static final String X_AXIS = "Year";
     private static final String Y_AXIS = "Count";
@@ -33,7 +33,7 @@ public class PatentOwnershipEvolutionQuery extends ChartQuery<Integer, Integer> 
     }
 
     @Override
-    public List<Pair<Integer, Integer>> get() {
+    public List<Pair<Long, Long>> get() {
         LOGGER.info("Running getPatentOwnershipEvolutionQuery search on " + ownersName + " owner and " + category + " category on " + DataSourceType.PATENT + " collection.");
 
         return queryCreator.getPatentOwnershipEvolutionQuery(DataSourceType.PATENT, ownersName, category);
